Tabla reusable 12%
