+++ src/Componentes/InventarioX/mock.js
@@ -1,6 +1,91 @@
-export const mock = [
+export const marcaactivo = [
     { label: 'Lenovo', value: 1 },
     { label: 'HP', value: 2 },
     { label: 'Dell', value: 3 },
     { label: 'Logitec', value: 4 }
 ]
+
+export const tipoactivo = [
+    { label: 'Teclado', value: 1 },
+    { label: 'Mouse', value: 2 },
+    { label: 'Portatil', value: 3 },
+    { label: 'Monitor', value: 4 }
+]
+
+export const Mockprueba = [
+    {
+        Tipo: "Teclado",
+        Serial: "ABCDEF123",
+        Marca: "Lenovo",
+        Modelo: "KeyBoard Pro 123"
+    },
+    {
+        Tipo: "Teclado",
+        Serial: "KB001",
+        Marca: "Logitech",
+        Modelo: "Ergo K860"
+    },
+    {
+        Tipo: "Monitor",
+        Serial: "MN002",
+        Marca: "Dell",
+        Modelo: "UltraSharp U2720Q"
+    },
+    {
+        Tipo: "Portátil",
+        Serial: "LT003",
+        Marca: "HP",
+        Modelo: "Pavilion 15"
+    },
+    {
+        Tipo: "Ratón",
+        Serial: "MO004",
+        Marca: "Razer",
+        Modelo: "DeathAdder Elite"
+    },
+    {
+        Tipo: "Impresora",
+        Serial: "PR005",
+        Marca: "Canon",
+        Modelo: "Pixma TR8520"
+    },
+    {
+        Tipo: "Teléfono móvil",
+        Serial: "PH006",
+        Marca: "Samsung",
+        Modelo: "Galaxy S21"
+    },
+    {
+        Tipo: "Disco duro externo",
+        Serial: "HD007",
+        Marca: "Seagate",
+        Modelo: "Backup Plus Slim"
+    },
+    {
+        Tipo: "Tablet",
+        Serial: "TB008",
+        Marca: "Apple",
+        Modelo: "iPad Air"
+    },
+    {
+        Tipo: "Altavoz Bluetooth",
+        Serial: "SP009",
+        Marca: "JBL",
+        Modelo: "Flip 5"
+    },
+    {
+        Tipo: "Cámara digital",
+        Serial: "CM010",
+        Marca: "Sony",
+        Modelo: "Alpha a7 III"
+    },
+    {
+        Tipo: "Auriculares inalámbricos",
+        Serial: "HD011",
+        Marca: "Bose",
+        Modelo: "QuietComfort 35 II"
+    }
+
+]
+
+export const headTableActivos = ["Tipo", "Serial", "Marca", "Modelo"]
