--- src/Componentes/InventarioX/tipoactivo.js
@@ -1,6 +1,0 @@
-export const tipoactivo = [
-    { label: 'Teclado', value: 1 },
-    { label: 'Mouse', value: 2 },
-    { label: 'Portatil', value: 3 },
-    { label: 'Monitor', value: 4 }
-]
