+++ src/Componentes/shared/table/table.jsx
@@ -1,15 +1,14 @@
 import { Row } from '../row/row';
 import { Rowhead } from '../RowHead/Rowhead';
-import { headTableUsuarios } from '../../Usuarios/mock'
 import './table.css'
 
-export const Table = () => {
+export const Table = ({headTable, dataTable}) => {
     return (
         <table className="table-container">
             <thead className='table-head'>
-                <Rowhead items={headTableUsuarios} />
+                <Rowhead items={headTable} />
             </thead>
-            <Row />
+            <Row elements={dataTable} />
         </table>
     )
 }
